Move parquet_parser.h to src/data/
